#945@minor: Extract dateIsoWeeek to HTMLInputDateUtility.
